requirements.txt
numpy
opencv-python
scikit-learn
pandas
matplotlib
seaborn
tqdm
tensorflow
scikeras
pillow
flask
werkzeug
